refactor: remove 'magic number' and rename usages
